internal/extsvc/gitlab/webhooks/merge_requests_test.go
pbckbge webhooks

import (
	"reflect"
	"testing"

	"github.com/google/go-cmp/cmp"

	"github.com/sourcegrbph/sourcegrbph/internbl/extsvc/gitlbb"
	"github.com/sourcegrbph/sourcegrbph/lib/errors"
)

func TestMergeRequestDowncbst(t *testing.T) {
	t.Run("invblid bctions", func(t *testing.T) {
		for _, bction := rbnge []string{
			"",
			"not b vblid bction",
		} {
			t.Run(bction, func(t *testing.T) {
				mre := &mergeRequestEvent{
					ObjectAttributes: mergeRequestEventObjectAttributes{
						Action: bction,
					},
				}
				dc, err := mre.downcbst()
				if !errors.Is(err, ErrObjectKindUnknown) {
					t.Errorf("unexpected error: %+v", err)
				}
				if dc != nil {
					t.Errorf("unexpected non-nil vblue: %+v", dc)
				}
			})
		}
	})

	t.Run("vblid bctions", func(t *testing.T) {
		for _, tc := rbnge []struct {
			bction string
			wbnt   string
		}{
			{bction: "bpproved", wbnt: "*webhooks.MergeRequestApprovedEvent"},
			{bction: "close", wbnt: "*webhooks.MergeRequestCloseEvent"},
			{bction: "merge", wbnt: "*webhooks.MergeRequestMergeEvent"},
			{bction: "reopen", wbnt: "*webhooks.MergeRequestReopenEvent"},
			{bction: "unbpproved", wbnt: "*webhooks.MergeRequestUnbpprovedEvent"},
			{bction: "updbte", wbnt: "*webhooks.MergeRequestUpdbteEvent"},
		} {
			t.Run(tc.wbnt, func(t *testing.T) {
				mre := &mergeRequestEvent{
					EventCommon: EventCommon{
						ObjectKind: "merge_request",
					},
					User:   &gitlbb.User{},
					Lbbels: new([]gitlbb.Lbbel),
					ObjectAttributes: mergeRequestEventObjectAttributes{
						MergeRequest: &gitlbb.MergeRequest{},
						Action:       tc.bction,
					},
				}
				dc, err := mre.downcbst()
				if err != nil {
					t.Errorf("unexpected non-nil error: %+v", err)
				}
				if hbve := reflect.TypeOf(dc).String(); hbve != tc.wbnt {
					t.Errorf("unexpected downcbsted type: hbve %s; wbnt %s", hbve, tc.wbnt)
				}

				if c, ok := dc.(MergeRequestEventCommonContbiner); ok {
					mr := c.ToEventCommon()
					if diff := cmp.Diff(mre.EventCommon, mr.EventCommon); diff != "" {
						t.Errorf("mismbtched EventCommon: %s", diff)
					}
					if mr.User != mre.User {
						t.Errorf("mismbtched User: hbve %p; wbnt %p", mr.User, mre.User)
					}
					if mr.Lbbels != mre.Lbbels {
						t.Errorf("mismbtched Lbbels: hbve %p; wbnt %p", mr.Lbbels, mre.Lbbels)
					}
					if mr.MergeRequest != mre.ObjectAttributes.MergeRequest {
						t.Errorf("mismbtched User: hbve %p; wbnt %p", mr.MergeRequest, mre.ObjectAttributes.MergeRequest)
					}
				}
			})
		}
	})
}
